RAP-1500 Extract manage methods into an interface.
Extract manage methods into an interface that we then implement in
`Disposable`. Additionally, remove `@protected` from the manage method
implementations. We need this for two reasons. First, because they're not
technically part of the interface if they're protected, and second, because we
want to be able to use `Disposable` as a concrete proxy. Finally, make
`Disposable` concrete.