RAP-1503 Implement Cache and ReferenceCache
Problem
---

The need to remove duplicates from or maintain a collection of instances
of a particular class demands the need for a Cache implementation. The
Cache should allow consumers to obtain a value from the cache or call an
underlying factory, explicitly put an item in the Cache, and remove
items from a Cache.

Solution
---

Implement the base Cache implementation. The get, getAsync, put, and
remove operations allow consumers to interact with the Cache as
described. Lifecycle methods and an update stream are provided allowing
consumers to extend the cache and implement custom behaviors or observe
the cache changes.

A specialization of the Cache is the ReferenceCache which maintains a
list of accesses from the Cache. Consumers can release a reference from
the cache which decrements the reference count. If the number of
releases equals the number of gets, the item is removed from the
ReferenceCache. This effectively allows consumers to memoize shared
resources.